requirements.txt
pytz
unidecode
BeautifulSoup
feedparser
gdata
python-dateutil
simplejson
Django>=1.4,<1.6
django-cms==2.4
